chore: move and rename the batching pre-execution layer
